+++ tests/trybuild/entity/type_comma_alone.rs
@@ -1,0 +1,5 @@
+use brood::Entity;
+
+type Entity = Entity!(,);
+
+fn main() {}