+++ tests/trybuild/entity/type_unexpected_token.rs
@@ -1,0 +1,8 @@
+use brood::Entity;
+
+struct A;
+struct B;
+
+type Entity = Entity!(A, + B,);
+
+fn main() {}