+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+### Added
+- `Entity!` macro for defining the type of an entity.
+- `reserve()` method on `World` for reserving capacity for additional entities made up of a specific set of components.
 
 ## 0.4.0 - 2022-12-03
 ### Added